Working Radio interaction between flight computer and controller

+++ lib/Navigation/PID.h
@@ -14,7 +14,7 @@ struct pid_state_t {
 
 class PID {
   public:
-    PID(float kP, float kI, float kD, int32_t maxOutput = 127, int32_t period = 0, int32_t maxIntegralComponent = 64)
+    PID(double kP, double kI, double kD, int32_t maxOutput = 127, int32_t period = 0, int32_t maxIntegralComponent = 64)
         : kP(kP), kI(kI), kD(kD), maxOutput(maxOutput), period(period), maxIntegral(maxIntegralComponent/kI) {}
 
     int32_t apply(int32_t input, int32_t target, int32_t intervalMicros, pid_state_t *sample = 0);
@@ -29,7 +29,7 @@ class PID {
     int32_t convertPeriodicError(int32_t e, int32_t p) { return ((e + (p - 1) / 2) % p + p) % p - (p - 1) / 2; }
 
     bool firstApply = true;
-    const float kP, kI, kD;
+    const double kP, kI, kD;
     const int32_t maxOutput, period;
     const int32_t maxIntegral;
     int32_t integral;
